hikptool: Added compilation options.
The following compilation options are added:

-Wuninitialized -Wundef -Wunused -Wdate-time
-Wshadow -Wvla -Wdisabled-optimization -Wempty-body -Wignored-qualifiers
-Wimplicit-fallthrough=3 -Wtype-limits -Wshift-negative-value -Wswitch-default
-Wframe-larger-than=8192 -Wshift-overflow=2 -Wwrite-strings -Wmissing-format-attribute
-Wformat-nonliteral -Wduplicated-cond -Wtrampolines -Wlogical-op -Wsuggest-attribute=format
-Wduplicated-branches -Wmissing-include-dirs -Wmissing-declarations
-Wreturn-local-addr -Wredundant-decls -Wfloat-conversion -Wmissing-prototypes
-Wno-discarded-qualifiers -Wstrict-prototypes

And alos rectify compilation alarms and errors.

+++ tool_lib/tool_lib.c
@@ -245,7 +245,7 @@ int tool_flock(const char *name, uint32_t operation, int *fd, const char *log_di
 	}
 
 	ret = snprintf(lock_file, sizeof(lock_file), "%s", log_dir);
-	if (ret < 0 || ret >= sizeof(lock_file)) {
+	if (ret < 0 || (size_t)ret >= sizeof(lock_file)) {
 		HIKP_ERROR_PRINT("generate flock [%s] folder name failed, errno is %d\n",
 				 log_dir, errno);
 		return -errno;
